modify file

+++ ABC/ABC_195/B_Many Oranges.py
@@ -3,8 +3,8 @@ import math
 a, b, k = map(int, input().split())
 k *= 1000
 
-mi = int(math.ceil(k / b))
-ma = int(math.floor(k / a))
+mi = math.ceil(k / b)
+ma = math.floor(k / a)
 
 if mi <= ma:
     print(mi, ma)
